pyproject.toml
[tool.poetry]
name = "whisperer"
version = "0.1.0"
description = "A command line utility for transcribing audio and video files"
authors = ["le_woudar <rollandkev@yahoo.fr>"]
license = "Apache-2.0"
readme = "README.md"

packages = [
    { include = "whisperer" }
]

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
click = "^8.1.7"
click-didyoumean = "^0.3.0"
shellingham = "^1.5.4"
rich = "^13.7.0"


[tool.poetry.group.lint.dependencies]
ruff = "^0.1.8"
bandit = "^1.7.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
commitizen = "^3.13.0"

[tool.poetry.scripts]
whp = "whisperer.main:cli"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = [
  "UP",  # pyupgrade
  "I",   # isort
  "S",   # flake8-bandit
  "B",   # flake8-bugbear
  "C90", # McCabe complexity
  "RUF"
]
ignore = ["S603"]

[tool.ruff.format]
quote-style = "single"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
